+++ src/tb/tb_dimension.cpp
@@ -28,7 +28,7 @@ void TBDimensionConverter::SetDPI(int src_dpi, int dst_dpi)
 void TBDimensionConverter::GetDstDPIFilename(const char *filename, TBTempBuffer *tempbuf) const
 {
 	int dot_pos = 0;
-	for (dot_pos = strlen(filename) - 1; dot_pos > 0; dot_pos--)
+	for (dot_pos = (int) strlen(filename) - 1; dot_pos > 0; dot_pos--)
 		if (filename[dot_pos] == '.')
 			break;
 	tempbuf->ResetAppendPos();
@@ -65,7 +65,7 @@ int TBDimensionConverter::GetPxFromString(const char *str, int def_value) const
 {
 	if (!str || !is_start_of_number(str))
 		return def_value;
-	const int len = strlen(str);
+	const int len = (int) strlen(str);
 	const int val = atoi(str);
 	if (len > 2 && strcmp(str + len - 2, "px") == 0)
 		return val;
@@ -79,7 +79,7 @@ float TBDimensionConverter::GetPxFromStringF(const char *str, float def_value) c
 {
 	if (!str || !is_start_of_number(str))
 		return def_value;
-	const int len = strlen(str);
+	const int len = (int) strlen(str);
 	const float val = (float) atof(str);
 	if (len > 2 && strcmp(str + len - 2, "px") == 0)
 		return val;
